Task 1: Added api/employee/{id}/structure endpoint
This endpoint facilitates returning ReportingStructure for the
 provided employeeId.

This required changes to the service & repository.  Here, I chose
to implement new endpoints rather than treat the missing property
(DirectReports) as a flaw in the original.  With further guidance
we could change the original, as the additive change would not
typically break callers.